backend/requirements.txt
anytree==2.8.0
appdirs==1.4.4
asgiref==3.2.7
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
distlib==0.3.0
distro==1.5.0
Django==3.0.7
django-filter==2.3.0
django-markdownx==3.0.1
djangorestframework==3.11.0
docutils==0.16
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
lxml==4.5.1
mallard-ducktype==1.0.2
Markdown==3.2.2
meson==0.54.2
msgpack==1.0.0
mysqlclient==2.0.1
ordered-set==3.1.1
packaging==20.4
pep517==0.8.2
Pillow==7.2.0
progress==1.5
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytoml==0.1.21
pytz==2020.1
ranger-fm==1.9.3
requests==2.23.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
webencodings==0.5.1